Settings Menu Escape Key Bug Fix
### Changes:
 - Added ability to, when user is in Main Menu, and Settings Menu is active, player can press Escape to exit Settings Menu
### Bug Fixes:
 - Fixed a bug where, if you press the 'Escape' key when in Settings Menu, it would toggle the Pause Menu or make the player visible and movable
   - Fixed by modifying TogglePauseCanvas() method in CanvasHandler.cs script to account for Settings Menu
### Known Bugs:
 - None
