+++ Test Building Mechanics/Assets/Scripts/Handlers/MainMenuButtonsHandler.cs
@@ -3,6 +3,9 @@ using UnityEngine.SceneManagement;
 
 public class MainMenuButtonsHandler : MonoBehaviour
 {
+    public CurrentKeybinds currentKeybindsScript;
+    public SettingsButtonsHandler settingsButtonsHandlerScript;
+
     public GameObject mainMenuCanvas;
     public GameObject settingsMenuCanvas;
 
@@ -21,6 +24,18 @@ public class MainMenuButtonsHandler : MonoBehaviour
     public void QuitGame()
     {
         Application.Quit();
+    }
+    #endregion
+
+    #region Back_Settings
+    void Update()
+    {
+        if (Input.GetKeyDown(currentKeybindsScript.pauseResumeKey) && (settingsMenuCanvas.activeSelf))
+        {
+            settingsButtonsHandlerScript.SaveAndApplyButton();
+            settingsMenuCanvas.SetActive(false);
+            mainMenuCanvas.SetActive(true);
+        }
     }
     #endregion
 }
